Add error handling, refactor source
